(fix) --tags will now properly skip tests if tags don't match
Because of how --tags was implemented, tests was still executed when they were supposed to be skippe. The result was just hidden.
This caused the output of tests that raised exception to still be showed. Aka they were not skipped.
New implementation makes it so we set tags on test definition with a decorator instead of as variabel in test.